requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pillow==10.2.0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
sympy==1.12
timm==0.9.16
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
